Remove the variants available? checking
The available? method was removed in spree. Fix the product page error
by removing the available checking. After the fix, all the photos of
variants is shown even it is out of stock. But it should be a better
behavour because the variants can still be selected by customer even it
is out of stock.
https://github.com/spree/spree_product_zoom/issues/21